+++ lib/ui/viewer/file_details/exif_item_widgets.dart
